tests/Node-insertBefore.html

<!DOCTYPE html>
<title>Node.insertBefore</title>
<script src="testharness.js"></script>
<div id="log"></div>
<script>
function testHRE(node) {
  test(function() {
    assert_throws("HIERARCHY_REQUEST_ERR", function() { node.insertBefore(document.createTextNode("fail"), null) })
    assert_throws("HIERARCHY_REQUEST_ERR", function() { node.insertBefore(null, null) })
  })
}
test(function() {
  var len = document.body.childNodes.length
  assert_throws("NOT_SUPPORTED_ERR", function() { document.body.insertBefore(null, null) }) // XXX test framework breaks
  assert_equals(document.body.childNodes.length, len)

  len = document.body.childNodes.length
  assert_throws("NOT_SUPPORTED_ERR", function() { document.body.insertBefore(null, document.body.firstChild) })
  assert_equals(document.body.childNodes.length, len)

  try {
    document.body.insertBefore({'a':'b'}, document.body.firstChild)
    assert_unreached()
  } catch (e) {
    assert_true(e instanceof TypeError)
  }

  document.body.setAttribute("abc", "pass")
  var attr = document.body.attributes[0]
  testHRE(attr)
  assert_equals(attr.value, "pass")
  assert_equals(attr.childNodes.length, 1)
  assert_equals(attr.childNodes[0].nodeValue, "pass")

  testHRE(document.createTextNode("Foo"))
})
</script>
